webscrapingtool/restapi/serializers.py
from .models import Outlet, Author
from rest_framework import serializers


class AuthorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Author
        fields = ('id', 'name', 'email')


class OutletSerializer(serializers.ModelSerializer):
    authors = AuthorSerializer(many=True, read_only=True)

    class Meta:
        model = Outlet
        fields = ('id', 'name', 'website', 'description', 'authors')
