requirements.txt
adium-theme-ubuntu==0.3.4
backports.csv==1.0.5
cli-helpers==0.2.3
click==6.7
configobj==5.0.6
Django==1.10.1
djangorestframework==3.4.7
humanize==0.5.1
pgcli==1.8.1
pgspecial==1.8.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
py==1.5.2
pycurl==7.43.0
Pygments==2.2.0
pytest==3.0.6
pytest-django==3.1.2
python-apt==1.1.0b1
setproctitle==1.1.10
six==1.11.0
sqlparse==0.2.4
terminaltables==3.1.0
unity-lens-photos==1.0
virtualenv==15.1.0
wcwidth==0.1.7
